requirements.txt
# =============================================================================
# WEBCRAFT PRO - PYTHON DEPENDENCIES
# =============================================================================

# Core FastAPI framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Image processing
Pillow==10.1.0

# HTTP requests
httpx==0.25.2
requests==2.31.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Email sending
aiosmtplib==3.0.1
email-validator==2.1.0

# File handling
aiofiles==23.2.1

# Date & Time
python-dateutil==2.8.2

# JSON handling
orjson==3.9.10

# Validation
validators==0.22.0

# Utils
slugify==0.0.1
python-slugify==8.0.1

# Background tasks (optional)
celery==5.3.4
redis==5.0.1

# Monitoring (optional)
psutil==5.9.6

# CORS
fastapi-cors==0.0.6